add number test
